trainer/requirements.txt
pyspark==3.3.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.2.2
pyarrow==12.0.0
fastparquet==2023.4.0
python-dotenv==1.0.0
pymongo==4.3.3
h5py==3.11.0 # joblib might use it, or for other potential model formats
# Removed: tensorflow, keras, bigdl-orca, torch, matplotlib, tqdm
